tests/submission/W3C/domhtml/HTMLStyleElement02.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<style media="screen" type="text/css"></style>
<title>NIST DOM HTML Test - STYLE</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/the-style-element.html#dom-style-media" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<p>Hello, World.</p>



<script>
var domtest = async_test("The media attribute identifies the intended medium of the style info.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vmedia;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("style");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vmedia = testNode.media;

      assert_equals(vmedia,"screen");
       
    domtest.done();
}));

</script>
<p>The media attribute identifies the intended medium of the style info.

    Retrieve the media attribute and examine its value.</p>
<div id="log"></div>
</body></html>
